refactor: centralize track feature module
