+++ src/components/workspace/DetailPanelContent.tsx
@@ -9,9 +9,7 @@ import { Badge } from "@/components/ui/badge";
 import { Textarea } from "@/components/ui/textarea";
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from "@/components/ui/accordion";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
-import { TrackVersions } from "@/components/tracks/TrackVersions";
-import { TrackStemsPanel } from "@/components/tracks/TrackStemsPanel";
-import { useTrackLike } from "@/hooks/useTrackLike";
+import { TrackVersions, TrackStemsPanel, useTrackLike } from "@/features/tracks";
 import { cn } from "@/lib/utils";
 import { StyleRecommendationsPanel } from "./StyleRecommendationsPanel";
 import type { StylePreset } from "@/types/styles";
