+++ src/features/tracks/components/TrackStemsPanel.tsx
