+++ src/features/tracks/ui/index.ts
@@ -1,0 +1,1 @@
+export { DetailPanel } from './DetailPanel';
